+++ NEWS.md
@@ -1,5 +1,6 @@
 # djprshiny 0.0.3.900
 * reactives can now be passed to plot functions through ... in plot_server
+* set font to Arial before downloading chart with djpr_plot_server
 
 # djprshiny 0.0.2.900
 * modifications to plot_server to enable nested modules