+++ NEWS.md
@@ -2,6 +2,7 @@
 * reactives can now be passed to plot functions through ... in plot_server
 * set font to Arial before downloading chart with djpr_plot_server
 * x_expand_mult argument added to djpr_ts_linechart()
+* tooltip column excluded from downloaded data
 
 # djprshiny 0.0.2.900
 * modifications to plot_server to enable nested modules